Renamed Alert.jsx to VhaJoinOrgAlert.jsx  and removed TODOs
